+++ tests/products/create_test.py
@@ -1,0 +1,15 @@
+
+# POST /products
+
+def test_create_product(client):
+    response = client.post('/products', data={
+        "aisle":"pending assignment",
+        "department": "pending assignment",
+        "name": "My New Product!",
+        "price": "100.00"})
+    response_body = response.get_json()
+    assert response.status_code == 200
+    assert isinstance(response_body, dict)
+
+# def test_auto_increment(arg):
+#     pass
