+++ tests/products/destroy_test.py
@@ -1,0 +1,2 @@
+
+# DELETE /products/:id
