+++ tests/products/update_test.py
@@ -1,0 +1,2 @@
+
+# PUT/POST /products
